Relocate service logic to external file and fix format
